pyproject.toml
[project]
name = "viper-dev"
version = "1.0"
description = "Code for ViPER: Foundational Vision for Your Robot with DINOv3"
authors = [
    { name = "Santiago Montiel-Marin" }
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "addict>=2.4.0",
    "hydra-core>=1.3.2",
    "jaxtyping>=0.3.2",
    "lightning==2.1.0",
    "matplotlib>=3.10.6",
    "numpy<2.0.0",
    "pillow>=11.3.0",
    "rootutils>=1.0.7",
    "setuptools>=77",
    "timm>=1.0.19",
    "torch==2.4.1",
    "torchvision>=0.19.1",
    "transformers>=4.56.0",
    "wandb>=0.21.3",
    "opencv-python>=4.11.0.86",
    "rich>=14.2.0",
    "segmentation-models-pytorch>=0.5.0",
]

[build-system]
requires = ["setuptools >= 77", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["viper"]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu124" },
]
torchvision = [
  { index = "pytorch-cu124" },
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv]
find-links = [
  "https://data.pyg.org/whl/torch-2.4.1+cu124.html"
]
